Flightlog improvements (#169)
* Fix form fields, no overlapping text and label

* Prevent duplicate flight crew

* Persons on board follows how many seats the airplane has and how big crew is selected

* Plus and minus buttons for setting number of landings

* Calculate new total time based on last known value

* Better UX for entering remaining fuel

* Revert "Merge pull request #162 from MIK-dev-team/dependabot/npm_and_yarn/npm_and_yarn-ef9012f86b"

This reverts commit 1b76cce4456f072be11fbe88e78c22080a19eb6f, reversing
changes made to 3f025bc1dda0c1c0c37febba49092bf50b87aa00.

+++ sql/schema/migration/V200__CreateViewFlightTimeTotals.sql
@@ -39,7 +39,7 @@ select
     format_flight_time(
         ajlb.minutes_at_start + l.sum_flight_mins
     ) as ac_total_flight_time,
-    ((ajlb.minutes_at_start + l.sum_flight_mins)/60)::int4 as ac_total_flight_hours
+    (ajlb.minutes_at_start + l.sum_flight_mins)/60.0 as ac_total_flight_hours
 
 
 from
